requirements.txt
PyQt5==5.15.9
Pillow==10.1.0
qrcode==7.4.2
opencv-python==4.9.0.80
numpy==1.26.4